requirements/dev.txt
# See more documentation in /docs/libraries.md

# Importing from base.txt
-r base.txt

# WSGI utility library with powerful debugger
# http://werkzeug.pocoo.org/
Werkzeug==0.11.2

# Debug information in a toolbar
# http://django-debug-toolbar.readthedocs.org/
django-debug-toolbar

# SSL localhost server
# https://github.com/teddziuba/django-sslserver
django-sslserver==0.15